refactor(#200): Split useWorkspaceManager tests into multiple files
- Split into core, composite, and persistence test files
- Remove eslint-disable max-lines
- Add shared testSetup file for localStorage mock
